Finish implementation of ListProjects
