+++ pkg/api/openapi/minder/v1/minder.swagger.json
@@ -1124,189 +1124,6 @@
         ]
       }
     },
-    "/api/v1/projects/availableRoles": {
-      "get": {
-        "operationId": "ProjectsService_ListRoles",
-        "responses": {
-          "200": {
-            "description": "A successful response.",
-            "schema": {
-              "$ref": "#/definitions/v1ListRolesResponse"
-            }
-          },
-          "default": {
-            "description": "An unexpected error response.",
-            "schema": {
-              "$ref": "#/definitions/googlerpcStatus"
-            }
-          }
-        },
-        "parameters": [
-          {
-            "name": "context.provider",
-            "description": "name of the provider",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          },
-          {
-            "name": "context.project",
-            "description": "ID of the project",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          },
-          {
-            "name": "context.retiredOrganization",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          }
-        ],
-        "tags": [
-          "ProjectsService"
-        ]
-      }
-    },
-    "/api/v1/projects/roles": {
-      "get": {
-        "operationId": "ProjectsService_ListRoleAssignments",
-        "responses": {
-          "200": {
-            "description": "A successful response.",
-            "schema": {
-              "$ref": "#/definitions/v1ListRoleAssignmentsResponse"
-            }
-          },
-          "default": {
-            "description": "An unexpected error response.",
-            "schema": {
-              "$ref": "#/definitions/googlerpcStatus"
-            }
-          }
-        },
-        "parameters": [
-          {
-            "name": "context.provider",
-            "description": "name of the provider",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          },
-          {
-            "name": "context.project",
-            "description": "ID of the project",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          },
-          {
-            "name": "context.retiredOrganization",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          }
-        ],
-        "tags": [
-          "ProjectsService"
-        ]
-      },
-      "post": {
-        "operationId": "ProjectsService_AssignRole",
-        "responses": {
-          "200": {
-            "description": "A successful response.",
-            "schema": {
-              "$ref": "#/definitions/v1AssignRoleResponse"
-            }
-          },
-          "default": {
-            "description": "An unexpected error response.",
-            "schema": {
-              "$ref": "#/definitions/googlerpcStatus"
-            }
-          }
-        },
-        "parameters": [
-          {
-            "name": "body",
-            "in": "body",
-            "required": true,
-            "schema": {
-              "$ref": "#/definitions/v1AssignRoleRequest"
-            }
-          }
-        ],
-        "tags": [
-          "ProjectsService"
-        ]
-      }
-    },
-    "/api/v1/projects/roles/{roleAssignment.subject}": {
-      "delete": {
-        "operationId": "ProjectsService_RemoveRole",
-        "responses": {
-          "200": {
-            "description": "A successful response.",
-            "schema": {
-              "$ref": "#/definitions/v1RemoveRoleResponse"
-            }
-          },
-          "default": {
-            "description": "An unexpected error response.",
-            "schema": {
-              "$ref": "#/definitions/googlerpcStatus"
-            }
-          }
-        },
-        "parameters": [
-          {
-            "name": "roleAssignment.subject",
-            "description": "subject is the subject to which the role is assigned.",
-            "in": "path",
-            "required": true,
-            "type": "string"
-          },
-          {
-            "name": "context.provider",
-            "description": "name of the provider",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          },
-          {
-            "name": "context.project",
-            "description": "ID of the project",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          },
-          {
-            "name": "context.retiredOrganization",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          },
-          {
-            "name": "roleAssignment.role",
-            "description": "role is the role that is assigned.",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          },
-          {
-            "name": "roleAssignment.project",
-            "description": "projectt is the projectt in which the role is assigned.",
-            "in": "query",
-            "required": false,
-            "type": "string"
-          }
-        ],
-        "tags": [
-          "ProjectsService"
-        ]
-      }
-    },
     "/api/v1/providers": {
       "get": {
         "operationId": "ProvidersService_ListProviders",
